+++ scripts/wiki/data-generators/evolutions.ts
@@ -14,11 +14,11 @@ import { SpeciesId } from "#enums/species-id";
 import { TimeOfDay } from "#enums/time-of-day";
 import { WeatherType } from "#enums/weather-type";
 import { getBiomeName } from "#utils/common";
-import { wikiSpeciesDataRegistry } from "./constants";
-import { writeWikiData } from "./helpers";
-import type { EvolutionWikiEntry } from "./types";
+import { wikiSpeciesDataRegistry } from "../constants";
+import { writeWikiData } from "../helpers";
+import type { EvolutionWikiEntry } from "../types";
 
-export function generateEvolutionsCsv(): void {
+export function generateEvolutionsData(): void {
   const entries: EvolutionWikiEntry[] = [];
 
   for (const speciesData of Object.values(wikiSpeciesDataRegistry.data)) {
